Fix C4244 narrowing warnings via type promotion
Promote function signatures and local variables to wider types to eliminate
MSVC C4244 warnings without losing precision:

- dwgbuffer: numRemainingBytes() -> int64_t, moveBitPos(int64_t),
  crc8() start/end -> int64_t, getUCSStr/get16bitStr size -> uint32_t
- dwgreader: maxPos -> uint64_t, startPos -> uint64_t,
  dwgSectionInfo compressed/encrypted -> uint64_t, pages map key -> uint64_t
- dwgutil: compressedSize/decompSize/compressedPos/decompPos -> uint64_t,
  compressedByte/decompByte index -> uint64_t, decompress18 locals -> uint64_t,
  decrypt18Hdr loop vars -> uint64_t, twoByteOffset litCount param -> uint64_t
- drw_entities/drw_objects: ms -> static_cast<uint32_t>(buf->size()),
  numRemainingBytes local -> int64_t
- dxfwriter: writeInt64 byte extraction -> static_cast<char>
- libdwgr/dwgreader15/18/21: static_cast<size_t> at allocator boundaries

Use static_cast only where narrowing is intentional per DWG spec
(e.g., 32-bit XOR mask, allocator boundaries on 32-bit platforms).

(cherry picked from commit 7f8b809e6a69a82f655c27efa855d131e98b4d6f)

+++ src/intern/dwgreader15.cpp
@@ -121,7 +121,7 @@ bool dwgReader15::readDwgHeader(DRW_Header& hdr){
         return false;
     if (!fileBuf->setPosition(si.address))
         return false;
-    std::vector<std::uint8_t> tmpByteStr(si.size);
+    std::vector<std::uint8_t> tmpByteStr(static_cast<size_t>(si.size));
     fileBuf->getBytes(tmpByteStr.data(), si.size);
     dwgBuffer buff(tmpByteStr.data(), si.size, &decoder);
     DRW_DBG("Header section sentinel= ");
